chore: Refactor file names Capitalize
