--- src/Controller/reviewController.ts
@@ -1,9 +1,0 @@
-import { IReview } from "../model/reviewModel";
-import ReviewService from "../service/reviewService";
-import GenericController from "./genericController";
-
-export default class ReviewController extends GenericController<IReview> {
-    constructor() {
-        super(new ReviewService())
-    }
-}
