+++ src/repository/implementations/ProductRepository.ts
@@ -1,6 +1,6 @@
-import { IProductFilter, IProductQuery, } from "../dtos/filterProduct";
-import Product, { IProduct } from "../model/productModel";
-import GenericRepository from "./genericRepository";
+import { IProductFilter, IProductQuery, } from "../../dtos/filterProduct";
+import Product, { IProduct } from "../../model/productModel";
+import GenericRepository from "./GenericRepository";
 
 
 export default class ProductRepository extends GenericRepository<IProduct> {
@@ -12,7 +12,7 @@ export default class ProductRepository extends GenericRepository<IProduct> {
         try {
 
             const query: Partial<IProductQuery> = {};
-            
+
             if (filter) {
                 if (filter.name) {
                     query.name = { $regex: `^${filter.name}|.*${filter.name}.*`, $options: "i" };
