+++ src/repository/implementations/ReviewRepository.ts
@@ -1,5 +1,5 @@
-import Review, { IReview } from "../model/reviewModel";
-import GenericRepository from "./genericRepository";
+import Review, { IReview } from "../../model/reviewModel";
+import GenericRepository from "./GenericRepository";
 
 
 export default class ReviewRepository extends GenericRepository<IReview> {
