--- src/service/categoryService.ts
@@ -1,9 +1,0 @@
-import { ICategory } from "../model/catagoryModel";
-import CategoryRepository from "../repository/categoryRepository";
-import GenericService from "./genericService"
-
-export default class CategoryService extends GenericService<ICategory> {
-    constructor() {
-        super(new CategoryRepository())
-    }
-}
